Migrated test for 'align_images_using_neighbors' out into its own test module
